Collection of new attributes in _Navigator and _Parentage all to help implement new _pathExistsBetween(arg) method bound to _Navigator.
The new _pathExistsBetween(arg) method returns true when a 'path' exists between self._client and arg. I chose the word 'path' here as a synonym for 'thread' because I think we may want to replace _isThreadable(arg) with _pathExistsBetween(arg) instead. The main idea behind _isThreadable(arg) is good but _isThreadable(arg) ignores a couple of important LilyPond behaviors. The easiest one of these can be seen in an example like this ...

  { \context Voice = 'foo' { ... } \context Voice = 'foo' { ... } }

... and this ...

  { \new Voice { ... } \new Voice { ... } }

... where we can call the first example 'consecutive *like-named* voices' and the second example 'consecutive *anonymous* voices. The threading model allows threading in both the first and second examples. But LilyPond allows threading in only the first example. (An easy test for LilyPond threading is, in both cases, to start a beam in the first voice and try to end it in the second; if the example renders without errors, then LilyPond allows threading between the two expressions, otherwise not.)

The logic in _pathExistsBetween(arg) conforms to LilyPond's ideas of what can and can not be threaded (with one small trivial example that I'll fix tomorrow). See core/test/test_path_exists_between.py for examples.