New _VoiceInterface implemented on _Component.
_VoiceInterface implements public creator, effective, explicit, name and signature properties. Explanations of all five properties follow even though probably only signature needs to remain public after testing finishes.

The read-only _VoiceInterface 'creator' property returns the component in the score hierarchy responsible for assigning voice to self._client. If not explicit Voice appears in the score hierarchy then it's likely that an explicit Staff, for example, assigns voice to notes and rests contained within. Where no score component assigns voice to self._client, returns None.

The read-only _VoiceInterface 'effective' property returns the voice that effectively governs self._client, or None.

The read-only _VoiceInterface 'explicit' property returns True when an explicit voice governs self._client, False when an implicit voice governs self._client, and None when no voice governs self._client.

The read-only _VoiceInterface 'signature' property returns a 1-tuple representation of the voice effectively governing self._client, or None when no voice governs self._client. The 1-tuple signature is of the form (name, ) when a named voice governs self._client; the tuple is of the form (id, ) when an anonymous voice governs self._client; the tuple is exactly equal to (-1, ) when the 'default' voice governs self._client.