Last work on _VoiceInterface for now.
Asking for t.voice.signature or t.voice.name works for a large number of Abjad structures, including all of the nonpathological cases. Some pathological cases are still problematic, all involving implicit LilyPond context instantiation.

We probably don't ever need to make *all* of the tests in voice/test/test_signature.py work. But if we do for some reason want to tackle the problem again, there are commented-out tests in test_signature.py that can serve as a starting point.

I'm moving forward with the spanned container code now after a full week of work on the voice-resolution stuff!