New _cappedDFSLeft( ) and _cappedDFSRight( ) methods implemented on _Navigator.
These two 'capped' depth-first search (DFS) methods return a generator which traverses self._client in depth-first order either left-to-right or right-to-left and which stop iteration after visiting all the nodes in self._client.

The 'capped' in the method names points to the fact that these methods 'cap' or limit iteration to self._client and children of self._client; iteration will never ascend to any parent of self._client.

(Compare with _depthFirstLeftToRight( ) and _depthFirstRightToLeft( ) which DO ascend to any parent of self._client and are therefore 'uncapped' or not limited in the same way.)