Simplified implementation of bound fracture( ) methods on _LeafSpannerAggregator.
Note that the public interface to _LeafSpannerAggregator.fracture( ) has changed slightly. The old interface was 

   t.spanners.fracture(
      classname = None, grob = None, attribute = None, value = None, 
      direction = 'both')

... and the new interface is now ...

   t.spanners.fracture(direction = 'both', classname = None)

... which takes away the ability to fracture spanners based on grob, attribute or value. So it is no longer possible to say t.spanners.fracture(attribute = 'color', value = 'red') directly. In the rare case that the functionality was being used, you can instead filter and fracture spanners by hand:

   for spanner in t.spanners.mine( ):
      if getattr(spanner, 'attribute', False) == 'color and \
         getattr(spanner, 'value', False) == 'red':
         spanner.fracture(spanner.index(t))

You can still fracture by class name using, for example, t.spanners.fracture('Tie') to fracture all tie spanners.

New leaf/spanner/test/fracture_test.py file with explicit examples of the 'receipt' that fracture( ) returns.

Also changed what happens when you fracture a spanner on either side of leaf t with a call to t.spanners.fracture('both') or its equivalent t.spanners.fracture( ). Such calls were previously fracturing spanners first to the left of t and then to the right of t and therefore returing two different 'receipts' of fracturation. This was confusing. Why should fracturing on either side of a leaf happen in two stages? And why should fracturing on the left happen before fracturing on the right instead of the other way around? Calls to fracture on both sides of leaf t now happen in a single step. The consequence of this is that calls to fracture( ) against the leaf spanner aggregator may now return a fracture( ) 'receipt' of either three or four elements. Three-element receipts are of the form (source, left, right) and return whenever you fracture either to the left or to the right of leaf t, but not both. Four-element receipts are of the form (source, left, center, right) and return whenever you fracture to both the left and right of leaf t at once.

Also in this commit, t.spanners.above( ), t.spanners.mine( ) and t.spanners.total( ) are now methods bound to _LeafSpannerAggregator rather than read-only properties of _LeafSpannerAggregator. This is to force us to be explicit about which level of the score tree we are querying when we ask _LeafSpannerAggregator to return spanners ... while still allowing each of these three methods to filter the set of spanners returned by either classname or any arbirary filter function named 'selector'.

Probably a good idea to run piece code and see if anything breaks since _LeafSpannerAggregator now implements a somewhat stingier public interface.