More _LeafSapnnerAggregator.fracture( ) simplifications.
Decided NOT to go with the idea that fracture( ) will cut 'up' and 'down' in the score structure. That is, calling t.spanners.fracture( ) on leaf t will fracture spanners attaching directly to t only.

Next step is to implement this same behavior on containers. This will be more of a chnage. Up to now containers have fractured 'down' by looking for spanners attaching to the first or last leaves in container. In the next revision or so, _ContainerSpannerAggregator will lose this ability.

Abjad lets you span whatever you want. Abjad lets you fracture whatever you want.

In both cases you must pick out the target(s) or spanning or fracturing explicitly.