Spanner insert( ) method now public.
Migrating the public interface of Spanner to implement (most of) the public interface of the built-in Python list. 

The Abjad Spanner already implements the following methods from the built-in Python list:

   * append( )
   * extend( )
   * index( )
   * insert( )
   * remove( )

The Abjad Spanner will soon implement the following additional parts of the public interface of the built-in Python list:

   * __getitem__
   * __setitem__
   * __delitem__
   * pop( )

The Abjad Spanner will NOT implement the following built-in list methods:

   * count( )
   * reverse( )
   * sort( )

The Abjad Spanner will preserve the following methods special to Spanner:

   * clear( )
   * fracture( )
   * fuse( )

The Abjad Spanner will deprecate the following methods special to Spanner:

   * appendleft( )
   * capture( )
   * extendleft( )
   * move( )
   * surrender( )

The changes will come over the next few revisions.