Spanner now implements the following:
   * __contains__
   * __delitem__
   * __getitem__
   * __len__
   * __setitem__
   
Slice assignment works the way you would expect.

NOTE: the methods all work on the *components* of p rather than the *leaves* of p. This represents a complete shift from the previous behavior of spanners only a few weeks ago in which iteration and length of spanners aliased only the leaves of spanners.

New test files added for each method.