New value-update pattern.
Implementation: _UpdateInterface implements new pattern for components to mark themselves (and the components in parentage) as '_current' or not. Back-tracking interfaces like _OffsetInterface can then take advantage of this up-to-dateness statefulness modelled by _UpdateInterface.

For example, _OffsetInterface now implements a new read-only 'flamingo' attribute as an optimized reimplementation of the existing 'score' attribute. (The 'flamingo' name is only temporary and should replace the existing 'score' attribute once this new _UpdateInterface code settles.) So you can now say t.offset.flamingo for any leaf t. At property-resolution time, the leaf's _OffsetInterface will check every component in the parentage of t. If any of the the components in the parentage of t are not current, then the _OffsetInterface of t calls _updateAll( ) on the _UpdateInterface of t. The _UpdateInterface of t then climbs all the way to score-level (or root-level) of the expression containing t and then searches depth-first through the entire expression containing t, updating the offset time of all leaves as it goes. The result is a fully update expression containing t, with all leaves from beginning to end now holding a correct value for score offset time. Following all this, the flamingo property returns. Crucially, on subsequent calls to flamingo offset, no expression-crawling is done and flamingo offset returns in O(1) time.

Test results, with ...

   v = Voice(Note(0, (1, 8)) * 1000)
   context2_timer = Timer('v[-1].offset.context2', 'from __main__ import v')
   flamingo_timer = Timer('v[-1].offset.flamingo', 'from __main__ import v')

... we get ...

>>> context2_timer.timeit(1)
0.68693208694458008
>>> context2_timer.timeit(5)
3.3768670558929443
>>> context2_timer.timeit(10)
6.792680025100708
>>> context2_timer.timeit(50)
33.890882015228271

... but with the value-update pattern we get ...

>>> flamingo_timer.timeit(1)
0.68927478790283203
>>> flamingo_timer.timeit(5)
0.00011301040649414062
>>> flamingo_timer.timeit(10)
0.00019478797912597656
>>> flamingo_timer.timeit(50)
0.00086617469787597656
>>> flamingo_timer.timeit(5000)
0.063992977142333984
>>> flamingo_timer.timeit(50000)
0.56457304954528809

... which is much better!

NOTE: music expressions become marked for update (that is, they become no longer current) when you call any of the following on any Abjad component:

   * _die( )
   * append( )
   * copy( )
   * embed( )
   * extend( )
   * insert( )
   * pop( )
   * remove( )

See the new update/test/test_update_interface.py file for examples.