_Parentage._iparentage deprecated.
Implementation: t._parentage._parentage now returns a list of all components in the parentage of t, INCLUDING t. In previous releases, t._parentage._parentage returned a list of all components in the parentage of t, excluding t.

If you want a list of all components in the parentage of t, excluding t, then ask for t._parentage._parentage[1:] instead.