New leaf bequath( ) method.
Extensions to the public interface: You can now say t.bequeath(Rest((1, 8))) or t.bequeath(FixedDurationTuplet((1, 8), Note(0, (1, 16)) * 3)) for for leaf t. Leaf t bequeaths position-in-spanners (if any) and position-in-parent (if any).

See the new leaf/test/test_leaf_bequeath.py testfile for examples.

Note that leaf bequeathal is essentially the second of the work implemented by leaf casting. Seen from the perspective of bequeathal, leaf casting does two things: 1. reproduces old, and 2. bequeaths from old to new. Since bequeath( ) now handles the second of these two steps, it makes sense to implement an independent function on leaves to simply copy leaf *without* spanners. In fact, it probably makes sense to implement an independent function on all components that copies the component in question without regard for the structural position of component; ie, copy and return a reference-pruned copy of component.