Changes to doc build system.
Changes to the public interface: the chapter.scr shell script has been deprecated and reimplemented as chapter.py in Python. So you should now say chapter.py instead of chapter.scr when building a chapter of the docs.

Extensions: the raw parser now finds the depth of the raw file being processed in relation to the root doc directory. This makes it possible to use classname aliases like <Note> in raw files of chapters at ALL levels of the doc directory. (This matters because chapters in the /helpers directory are at a lower level of the directory than chapters in the /chapters directory; previous doc site code assumed that all chapters were at the level of the /chapters directory, which is no longer true.)

Also, the raw parser now finds ABJADPATH from abjad.cfg.cfg and determines LilyPond version number dynamically.