New Abjad Function class to model the names of Scheme functions known to LilyPond.
Extensions to the public interface: you can now say t.meter.break_visibility = Function('end-of-line-invisible') to get:

   \new Staff \with {
      \override TimeSignature #'break-visibility = #end-of-line-invisible
   } {
      c'8
      d'8
      e'8
      f'8
   }

The key point is that end-of-line-invisible is the name of a Scheme function known to LilyPond and that end-of-line-invisible must format with a preceeding octothorpe but with no preceeding apostropher.

Implementation details: parser modified to know about the knew Function class.

See scm/test/test_scheme_function.py for an example.

Note that we should probably also implement a Vector class to model Scheme vectors. This will allow us transform Python [True, True, False] to LilyPond '#(#t #t #f).