Spanned leaf casting bug fix.
Implementation details: changed one line in _transfer_all_attributes( ) to make sure that an *iterable (tuple or list) of spanners* gets passed to _ComponentSpannerAggregator._update( ). Passing a *single spanner* to _ComponentSpannerAggregator._update( ) causes the atomic elements of that spanner to add to _ComponentSpannerAggregator._spanners, which is bad. Better is to pass an iterable containing a single spanner so that the *entire spanner* adds to _ComponentSpannerAggregator._spanners.

New leaf/test/test_leaf_cast_spanned.py testfile to catch regression.