New RigidMeasure.
Extensions to the public interface:

1. New RigidMeasure class exists. You can now say, for example, t = RigidMeasure((4, 8), scale(4)); t will then raise an UnderfullMeasureException at format-time in the case that t.meter.effective.duration does not equal t.duration.contents. See measure/test/test_measure_rigid.py for examples.

2. New _Measure.full predicate. Returns True when t.meter.effective.duration equals t.duration.contents, and False otherwise.

3. New MeasuresMisdurated check.

There is some weirdness with how the f( ) helper is working. An underfull RigidMeasure raises an UnderfullMeasureException at format-time, which is a good thing. One unintended consequence of this is that hasattr(t, 'format') returns *False* when t.format raises an exception. This means that an underfull RigidMeasure t currently raises an UnderfullMeasureException when you call t.format, but raises no exception when you call f(t) instead. Inelegancy but doesn't prevent us from using RigidMeasure for a while.

Going forward, the intent is to rename the current Measure class to ProlatingMeasure and to rename RigidMeasure to Measure. Try out RigidMeasure for a day or two and report any weirdness. Will swap names soon.