Changes to tcopy( ) helper.
Details: tcopy( ) is now somewhat 'dynamic' in the sense of DynamicMeasure. Previously, if you had ...

   t = FixedDurationTuplet((2, 8), scale(3))

... and you took tcopy(t[:2]), you wound up with a new FixedDurationTuplet with target duration set equal to Rational(2, 8), just as is the case with t. This meant that the output of tcopy, in this case, was a trivial 1:1 tuplet.

Now when you take tcopy(t[:2]) you get a new FixedDurationTuplet with a different target duration but with *the same multiplier*, ie Rational(2, 3). That is, the output of tcopy( ) 'shrinks' or 'trims' as necessary.

Additionally, there's another bit of code in tcopy( ) that looks for the denominator of t, if it exists, and tries to render output in terms of that denominator. This 'denominator preservation' applies, of course, only to measures.