New measures_subsume( ) helper.
Extensions to the public interface: you can now say measures_subsume(expr) to iterate all specially formed RigidMeasures of length 1 that contain only a tuplet. The result will change the meter of t, strip the contents of t from their containing measure, and, where necessary, adjust the written duration of the leaves in t appropriately. Example:

   t = RigidMeasure((6, 8), [FixedDurationTuplet((6, 8), scale(7))])
   measures_subsume(t)

      \time 21/28
      \scaleDurations #'(4 . 7) {
         c'8.
         d'8.
         e'8.
         f'8.
         g'8.
         a'8.
         b'8.
      }

This example subsumes a 7:6 tuplet, changing the meter of t from 6/8 to 21/28 and adjusting the written duration of the leaves in t from 1/8 to 3/16, ie, adding a dot to each of the leaves in t.

See helpers/test/test_measures_subsume.py for further examples.