New splice_after( ) helper.
Extensions to the public interface: you can now say splice_after(t, new_components) for any Abjad component t. The helper finds the position of t in any enclosing score and inserts each of the new components in new_components in the score immediately after t. The helper then iterates each of the spanners attached to t; for each spanner, the helper finds the index of t in the spanner and insert each of the new components in new_components in the spanner. So splice_after( ) preserves both parentage and spanners. The helper differs from, for example, list.append( ), list.insert( ), list.extend( ) in that the splice_after( ) operates on a sibling component rather than a parent, conatining component.

See helpers/test/test_splice_after.py for examples.