New flip( ) helper.
Extensions to the public interface: you can now say flip(expr) for any expression expr. The helper flips expr one place to the right in the parent of expr, if parent exists. The helper flips expr place to the right in any spanners attached to expr and to the successor of expr. Example:

   t = Voice(scale(4))
   Beam(t[:2])
   Beam(t[2:])

   \new Voice {
      c'8 [
      d'8 ]
      e'8 [
      f'8 ]
   }

   flip(t[1])

   \new Voice {
      c'8 [
      e'8 ]
      d'8 [
      f'8 ]
   }

The example flips d'8 and e'8 and leaves preserves the structure of both spanners.

See helpers/test/test_flip.py for further examples.