New private t.parentage._splice( ) method.
Implementation details: if is now possible for developers to say t.parentage._splice(components) for any Abjad component t. If t has a parent, the method inserts all components in the parent of t, immediately after t. If t has no parent, the method returns None. Because parentage._splice( ) doesn't handle spanners, it is possible splice components into the parent of t and render spanners attached to t discotinguous at the same time; this is also the reason that the method is private. Example:

   notes = scale(4)
   first = notes[:2]
   t = Voice(first)
   Beam(t[:])

   \new Voice {
      c'8 [
      d'8 ]
   }

   last = notes[2:]
   t[1].parentage._splice(last)

   \new Voice {
      c'8 [
      d'8 ]
      e'8
      f'8
   }

See parentage/test/test_parentage_splice.py for this example.