New container_shatter( ) helper.
Extensions to the public interface: you can now say container_shatter(t) for any Abjad container t. The helps wraps each of the top-level elements in t in a new RigidMeasure, and bequeaths the in-score and in-spanners position of t to the sequence of mesaures newly created. Example:

   t = Voice([RigidMeasure((3, 8), scale(3))])
   Beam(t[0])
   container_shatter(t[0])

   \new Voice {
         \time 1/8
         c'8 [ ]
         \time 1/8
         d'8 [ ]
         \time 1/8
         e'8 [ ]
   }

See helpers/test/test_container_shatter.py for two further examples.