New container_glom_by_count( ) helper.
Signature:

   container_glom_by_count(container, counts, target = Note(0, (1, 4)), 
      direction = 'big-endian')

Extensions to the public interface: you can now say container_glom_by_count(container, counts, **kwargs), for any sequential Abjad container. The helper gathers the elements in container together in numbers that are equal to the elements in counts (by calling container_partition_by_counts( ) first). The helper then creates new notes or rests, as specified by the 'target' keyword, and repopulates container with the newly created notes of rests. Note that any pitch material assigned to any notes or chords in container is lost in this process; this means that the helper is best used during pure rhythmic construction.

Example:

   t = Voice(run(5))
   Beam(t)
   container_glom_by_count(t, [1, 2, 2])

   \new Voice {
      c'8 [
      c'4
      c'4 ]
   }

Note that spanner handling is not yet implemented. This means that it is currently best to beam the container whose contents are to be glommed rather than beaming the leaves that are to be glommed.

See helpers/test/test_container_glom_by_count.py for further examples.