New measure_scale_and_remeter(measure, multiplier) helper.
Extensions to the public interface: you can now say measure_scale_and_remeter(measure, multiplier) for any RigidMeasure measure and any positive Rational multiplier. The helper multiplies the duration of every element in measure by multiplier. The helper then rewrites the meter of measure as appropriate.

Note: the algorithm is complex in that it incorporates a critical set of composer heuristics for measure scaling. It works surprisingly well, moving between dotted and nondotted noteheads and between binary and nonbinary meters when you would expect it to, and in the right ways.

Example:

      t = RigidMeasure((3, 8), scale(3))
      measure_scale_and_remeter(t, Rational(2, 3))

           \time 3/12
           \scaleDurations #'(2 . 3) {
                   c'8
                   d'8
                   e'8
           }

See helpers/test/test_measure_scale_and_remeter.py for this example and five others.

Note too that the name measure_scale_and_remeter( ) is probably provisional and will change to something shorter eventually.