Renamed _SpannerReceptor.spannedAbove to _SpannerReceptor.parented to comply with pattern of one-word public attributes.
Only the MetricGrid spanner and leaf_scale( ) helper had to change to accommodate the new name of the property.

So the thing to remember is that there's no the idea of a spanner receptor being 'parented' by a spanner: a spanner receptor, like t.beam, is 'parented' if and only if a spanner attaches to any of the components in the parentage of t, including t.