New _SpannerReceptor.chain property.
The _SpannerReceptor.chain property returns an (immutable) tuple of all leaves in spanner, if client of spanner receptor is spanned. In the case that the client of spanner receptor *is not* spanned, then _SpannerReceptor.chain returns an (immutable) 1-tuple of client only.

Intended primarily to model the idea of a 'tie chain'. According to the idea of a tie chain, untied notes return simply as (Note(0, (1, 4)), ) but tied notes return as (Note(0, (1, 8)), Note(0, (1, 8)), Note(0, (1, 8))), for example.

See spanner/test/test_spanner_receptor_chain.py for examples.