New Spanner.written attribute.
Extensions to the public interface. You can now say p.written to return the sum of the written duration of all leaves in any spanner p. Compare to p.duration which has existed for a very long time and which returns the sum of the prolated duration of all leaves in p. 

New test files:

   spanner/test/test_spanner_duration_prolated.py
   spanner/test/test_spanner_duration_written.py

Intended use of p.written is to model the written duration of a tie chain. This is important when, for example, you want to replace a tie chain with a tuplet, which will show up in a future commit.

Note that now that there is more than one spanner duration attribute, it probably makes sense to create a _SpannerDurationInterface to house p.duration.written and p.duration.prolated. This will show up in a later commit sometime.