New divide_leaf(leaf, divisions, prolation_type) helper.
Extensions to the public interface: you can now say divide_leaf(leaf, divisions = 2, prolation_type = 'diminution') for any leaf t. The helper bequeaths the score position of leaf with a newly created fixed-duration tuplet, such that the target duration of the newly created tuplet equals the written duration of t, and such that the newly created tuplet contains 'divisions' number of evenly spaced notes. The prolation keyword can be set to either 'diminution' or 'augmentation' and is used only in those cases where there exist two different ways of notating the same tuplet ratio, for example, to choose between 6:5 and 3:5, the first of which is a diminution and the second of which is an augmentation.

   t = divide_leaf(Note(0, (3, 16)), 5, 'diminution')
   \fraction \times 3/5 {
      c'16
      c'16
      c'16
      c'16
      c'16
   }

   t = divide_leaf(Note(0, (3, 16)), 5, 'diminution')
   \fraction \times 6/5 {
      c'32
      c'32
      c'32
      c'32
      c'32
   }

See helpers/test/test_divide_leaf.py for further examples.