New _get_attached_spanners(component_list) private helper.
Return all spanners attaching to any of the components in component_list.

   t = Staff(scale(4))
   b1 = Beam(t[:2])
   b2 = Beam(t[2:])
   crescendo = Crescendo(t)

   \new Staff {
      c'8 [ \<
      d'8 ]
      e'8 [
      f'8 ] \!
   }

   spanners = _get_attached_spanners(t[:])

   assert b1 in spanners
   assert b2 in spanners
   assert crescendo not in spanners

Note that crescendo is not in spanners because crescendo attaches to t rather than to any of the components passed in to _get_attached_spanners.

See heleprs/test/test_get_attached_spanners.py for this and two further examples.