New bequeath_multiple(old_components, new_components) helper.
Extensions to the public interface: you can now say bequeath_multiple(old_components, new_components) for any list of successive Abjad components, old_components, and any list of orphan Abjad components, new_components. The helper bequeaths the position-in-score of old_components to new_components; the helper also bequeaths the position-in-spanners of old_components to new_components.

Note that this helper is quite powerful. There's spanner magic going on in the implementation here that relies on the idea of 'dominant' spanners. The code reads easily enough. But the best way to get a feeling of the spanner behavior is to look at all examples in the test file in order.

Example:

   t = Staff(scale(4))
   b1 = Beam(t[:2])
   b2 = Beam(t[2:])
   crescendo = Crescendo(t[:])

   \new Staff {
      c'8 [ \<
      d'8 ]
      e'8 [
      f'8 ] \!
   }

   new_notes = Note(12, (1, 16)) * 5
   bequeath_multiple(t[1:3], new_notes)

   \new Staff {
      c'8 [ ] \<
      c''16
      c''16
      c''16
      c''16
      c''16
      f'8 [ ] \!
   }

This example bequeaths the position of the middle two notes in staff to a run of five new sixteenth notes. Because the crescendo 'dominates' these middle two notes, the crescendo grows to include all five new notes of the run. On the other hand, neither the first nor second beam 'dominates' these middle two notes; and, for that reason, the first beam shrinks to encompass only c'8 and the last beam shrinks to encompass only f'8.

See helpers/test/test_bequeath_multiple.py for this and four further examples. It is recommended to read all five examples in the test file and observe spanner behavior.