New divide_tie_chain(tie_chain, divisions, prolation_type) helper.
Extensions to the public interface: you can now say divide_tie_chain(tie_chain, divisions, prolation_type) for any Abjad tie chain, tie_chain. The helper creates a new fixed-duration tuplet such that the target duration of the tuplet equals the written duration of the tie chain, and such that the number of notes in the tuplet equals divisions.

This helper is a particularly powerful generalization of divide_leaf( ). Note that spanners are handled in a powerful way, preserving dominant spanners.

Example:

   t = Staff(construct.notes(0, [(5, 32), (5, 32)]))
   Beam(t[:2])
   Beam(t[2:])
   Crescendo(t[:])

   \new Staff {
      c'8 [ \< ~
      c'32 ]
      c'8 [ ~
      c'32 ] \!
   }

   divide_tie_chain(t[0].tie.chain, divisions = 6)

   \new Staff {
      \fraction \times 5/6 {
         c'32 [ \<
         c'32
         c'32
         c'32
         c'32
         c'32 ]
      }
      c'8 [ ~
      c'32 ] \!
   }

See helpers/test/test_divide_tie_chain.py for this and one other example.