New measures_project(expr) public helper.
Extensions to the public interface: you can now say measures_project(expr) for any Abjad expression. The helper iterates expr. For every measure in expr, if measure is nonbinary and able to rewrite as binary, the helper rewrites measure as binary.

Example:

   inner = FixedDurationTuplet((2, 16), run(3, Rational(1, 16)))
   notes = run(2)
   outer = FixedDurationTuplet((2, 8), [inner] + notes)
   t = RigidMeasure((2, 8), [outer])
   diatonicize(t)
   measures_subsume(t)

      \time 3/12
      \scaleDurations #'(2 . 3) {
         \times 2/3 {
            c'16
            d'16
            e'16
         }
         f'8
         g'8
      }

   measures_project(t)

      \time 2/8
      \times 2/3 {
         \times 2/3 {
            c'16
            d'16
            e'16
         }
         f'8
         g'8
      }

Here a measure in 3/12 projects a measure in 2/8.