New class-level 'color' setting on FixedDurationTuplet and FixedMultiplierTuplet for printing the brackets and number of trivial tuplets and coloring them blue.
Note that this setting is made in an unusual way!

Example:

   FixedDurationTuplet.color = True
   t = FixedDurationTuplet((3, 8), scale(3))

   \tweak #'color #blue
   \times 1/1 {
      c'8
      d'8
      e'8
   }

The settings is made *on the CLASS* and not *on the instance*.

The intended use of the setting is to enable visual location of trivial tuplets at composition-time. It is further intended that 'trivial tuplet coloring' will be turned off at final print-time of the score. Consider this is a visualization and debugging feature rather than a professional engraving feature.

See tuplet/fd/test/test_fdtuplet_color.py for this and one other example.