New _components_likely_multiplier(components) private helper.
The helper looks at all tie chains in components list and checks to see if they all exhibit the same (non-assignable) factor in the numerator of their written durations. The intent of the function is to help undo measure subsumption.

   t = Staff(scale(4))
   container_contents_scale(t, Rational(5, 4))
   assert _components_likely_multiplier(t[:]) == Rational(5, 4)

See helpers/test/test_components_likely_multiplier.py for this and five other examples.