container_contents_scale(container) is now smart enough to 'undo' certain (non-assignable) tied note values.
   t = Voice(scale(4))
   container_contents_scale(t, Rational(5, 4))

   \new Voice {
      c'8 ~
      c'32
      d'8 ~
      d'32
      e'8 ~
      e'32
      f'8 ~
      f'32
   }

   container_contents_scale(t, Rational(4, 5))

   \new Voice {
      c'8
      d'8
      e'8
      f'8
   }

In this example, the first call to container_contents_scale( ) with a multiplier of 5/4 creates many tied values. However, the second call to container_contents_scale( ) with the multiplicative inverse of 4/5 'undoes' the first call and leaves only untied values.

See helpers/test/test_container_contents_scale.py for this new example.