Public meter_make_binary(meter, contents_multiplier = Rational(1)) now takes an optional contents_multiplier argument. Explanation follows.
The helper previously took no keyword argument and would convert nonbinary meters like Meter(15, 24) into binary meters like Meter(10, 16). That conversion from 15/24 to 10/16 is fine. But sometimes it's desirable to go from 15/24 to 5/8.

What determines whether we want 15/24 -> 10/16 or instead 15/24 -> 5/8 is any 'contents multiplier' that might likely apply to the measure which 15/24 governs.

So how do we know if there's a likely contents multiplier in action on the contents of the measure with 15/24 governs? Well, there's a new private _contents_likely_multiplier( ) helper that will reveal that answer. That function is heuristic and will probably have to be improved over the next couple of weeks as it gets use. But it does a reasonably good job of, for example, noticing that a measure is full of a bunch of values like [5/32, 5/32, 5/32] all expressed as tie chains and flagging the contents of that measure as very likely coming from [1/8, 1/8, 1/8] all scaled by 5/4 to cause augmentation.

So, the way to use this new option 'contents_multiplier' argument to meter_make_binary is to pipe the output of _contents_likely_multiplier in to it.

Here are contrastive examples:

   Make n/24 meters into n/16 meters, where possible.

   assert meter_make_binary(Meter(1, 24)) == Meter(1, 24)
   assert meter_make_binary(Meter(2, 24)) == Meter(2, 24)
   assert meter_make_binary(Meter(3, 24)) == Meter(2, 16)
   assert meter_make_binary(Meter(4, 24)) == Meter(4, 24)
   assert meter_make_binary(Meter(5, 24)) == Meter(5, 24)
   assert meter_make_binary(Meter(6, 24)) == Meter(4, 16)
   assert meter_make_binary(Meter(7, 24)) == Meter(7, 24)
   assert meter_make_binary(Meter(8, 24)) == Meter(8, 24)

   Make n/24 meters into n/8 meters, where possible.

   assert meter_make_binary(Meter(1, 24), Rational(99)) == Meter(1, 24)
   assert meter_make_binary(Meter(2, 24), Rational(99)) == Meter(2, 24)
   assert meter_make_binary(Meter(3, 24), Rational(99)) == Meter(1, 8)
   assert meter_make_binary(Meter(4, 24), Rational(99)) == Meter(4, 24)
   assert meter_make_binary(Meter(5, 24), Rational(99)) == Meter(5, 24)
   assert meter_make_binary(Meter(6, 24), Rational(99)) == Meter(2, 8)
   assert meter_make_binary(Meter(7, 24), Rational(99)) == Meter(7, 24)
   assert meter_make_binary(Meter(8, 24), Rational(99)) == Meter(8, 24)

Notice that the *value* of 'contents_multiplier' doesn't matter; it need only *not equal* 1 to trigger one less power of two than would otherwise apply.