mathtools tools module now available. This module will house general math helper functions. Strictly math helper functions currently under abjad/helpers such as factors( ) should live heres. The module currently contains the following new helpers:
divide(total, parts):   returns a list of fractions of  the given total with the proportions found in the given parts. This function is analogous to the divide( ) helper function currently under abjad/helpers/, but operates on, and returns,   scalars, not Abjad objects.

fragment(total, cell):  fragments the given total value into the fractions found in the cell list. sum(cell) must be <= total. The remainder is appended to the given cell. e.g.
>>> fragment(Rational(1,2), [(1, 6), (1, 10)])
[Rational(1, 6), Rational(1, 10), Rational(7, 30)]