list_helpers(search_string) now accepts a more general search_string as input, rather than the less general start_string, as before.
Example:

   abjad> list_helpers('x')

   Abjad r1185 implements 3 helpers containing the string 'x':

      cyclic_extend, excise, tuplet_contents_fix

Also, list_helpers( ) no longer raises an IndexError when list_helpers( ) finds no helpers containing search_string.

Example:

   abjad> list_helpers('foo')

   Abjad r1185 implements 0 helpers containing the string 'foo'.

Note that there are still no tests for list_helpers(search_string) because the helper prints information but returns no values. And because the information that the helper prints depends strongly on the revision of Abjad that is running.