New _Parentage._containment private attribute. New _ContextContainmentSignature class.
You can now ask for t.parentage._containment for any Abjad component t. A _ContextContainmentSignature instance returns. The _ContextContainmentSignature identifies the root and first voice, staff and score in the parentage of t. Notice that the _ContextContainmentSignature gives *the first* voice, staff and score in the parentage of t rather than, say, *all* voices, staves or scores in the parentage of t. This is important ecause the parentage of an Abjad component may potentially contain many (nest) voices. _ContextContainmentSignature captures *the first* voice, staff and score in the parentage of t because it appears that this concept of the first (or 'nearest') enclosing context is what matters for voice-resolution logic.

The intention behind _Parentage._containment is to allow for an immediate test of whether a reference pathway exists between any two Abjad components. This sort of logic is currently implemented in _Navigator._pathExistsBetween(arg). But it may be possible to deprecate all of the _Navigator._pathExistsBetween(arg) code in favor of the much simpler _Parentage._containment code.

Here's an example of context containment signature being used to assert pathways between different Abjad components:

   t = Voice(run(4))
   t.invocation.name = 'foo'
   t.insert(2, Parallel(Voice(run(2)) * 2))
   t[2][0].invocation.name = 'foo'
   diatonicize(t)
   t.notehead.color = 'red'



   \context Voice = "foo" \with {
           \override NoteHead #'color = #red
   } {
           c'8
           d'8
           <<
                   \context Voice = "foo" {
                           e'8
                           f'8
                   }
                   \new Voice {
                           g'8
                           a'8
                   }
           >>
           b'8
           c''8
   }

   signatures = [leaf.parentage._containment for leaf in t.leaves]

   signatures[0] == signatures[1]
   signatures[0] == signatures[2]
   signatures[0] != signatures[4]
   signatures[0] == signatures[6]

   signatures[2] == signatures[0]
   signatures[2] == signatures[3]
   signatures[2] == signatures[4]
   signatures[2] == signatures[6]

   signatures[4] != signatures[0]
   signatures[4] != signatures[2]
   signatures[4] == signatures[5]
   signatures[4] == signatures[6]

See parentage/test/test_parentage_containment.py for this and eight other substantial examples.